hospinet/requirements.txt
django>=1.5
django_countries
django-treemenus
sorl-thumbnail
django-tastypie
django-extensions
south
django-templated-email
django-userena
django-bootstrap-pagination
django-crispy-forms
django-select2-forms
django-constance[database]